+++ packages/layer-composer/src/generators/types.ts
@@ -236,6 +236,7 @@ export interface VesselEventsGeneratorConfig extends GeneratorConfig {
   data: RawEvent[]
   color?: string
   track?: TrackGeneratorConfigData
+  showIcons?: boolean
   currentEventId?: string
 }
 
@@ -356,6 +357,8 @@ export type Ruler = {
   isNew?: boolean
 }
 
+export type HeatmapAnimatedInteractionType = 'presence' | 'viirs' | 'fishing-effort'
+
 export interface HeatmapAnimatedGeneratorSublayer {
   id: string
   datasets: string[]
@@ -365,6 +368,7 @@ export interface HeatmapAnimatedGeneratorSublayer {
   visible?: boolean
   breaks?: number[]
   legend?: GeneratorLegend
+  interactionType?: HeatmapAnimatedInteractionType
 }
 
 // ---- Heatmap Generator color ramps types