+++ packages/layer-composer/src/generators/vessel-events/vessel-events.ts
@@ -14,7 +14,7 @@ import {
   getVesselEventsGeojson,
   getVesselSegmentsGeojson,
   filterGeojsonByTimerange,
-  setActiveEvent,
+  filterFeaturesByTimerange,
 } from './vessel-events.utils'
 
 interface VesselsEventsSource extends GeoJSONSourceRaw {
@@ -34,7 +34,7 @@ class VesselsEventsGenerator {
   }
 
   _getStyleSources = (config: GlobalVesselEventsGeneratorConfig): VesselsEventsSource[] => {
-    const { id, data, track, start, end, currentEventId } = config
+    const { id, data, track, start, end } = config
 
     if (!data) {
       // console.warn(`${VESSEL_EVENTS_TYPE} source generator needs geojson data`, config)
@@ -42,24 +42,16 @@ class VesselsEventsGenerator {
     }
 
     const geojson = memoizeCache[config.id].getVesselEventsGeojson(data) as FeatureCollection
-    const newData: FeatureCollection = setActiveEvent(geojson, config.currentEventId || null)
-
-    if (config.start && config.end) {
-      const startMs = new Date(config.start).getTime()
-      const endMs = new Date(config.end).getTime()
-      newData.features = newData.features.filter((feature) => {
-        return (
-          feature.properties &&
-          feature.properties.timestamp > startMs &&
-          feature.properties.timestamp < endMs
-        )
-      })
-    }
+    const featuresFiltered = memoizeCache[config.id].filterFeaturesByTimerange(
+      geojson.features,
+      config.start,
+      config.end
+    )
 
     const pointsSource: VesselsEventsSource = {
       id: `${id}_points`,
       type: 'geojson',
-      data: newData,
+      data: { ...geojson, features: featuresFiltered },
     }
     const showTrackSegments = this._showTrackSegments(config)
 
@@ -91,7 +83,7 @@ class VesselsEventsGenerator {
 
     const activeFilter = ['case', ['==', ['get', 'id'], config.currentEventId || null]]
 
-    const pointsLayers = [
+    const pointsLayers: (CircleLayer | SymbolLayer)[] = [
       {
         id: `${config.id}_background`,
         type: 'circle',
@@ -113,7 +105,11 @@ class VesselsEventsGenerator {
           generatorId: config.id,
         },
       } as CircleLayer,
-      {
+    ]
+
+    const showIcons = config.showIcons !== undefined ? config.showIcons : true
+    if (showIcons) {
+      pointsLayers.push({
         id: `${config.id}_outline`,
         source: `${config.id}_points`,
         ...(showTrackSegments && { maxzoom: POINTS_TO_SEGMENTS_ZOOM_LEVEL_SWITCH }),
@@ -126,8 +122,8 @@ class VesselsEventsGenerator {
         metadata: {
           group: Group.Point,
         },
-      } as SymbolLayer,
-    ]
+      } as SymbolLayer)
+    }
 
     if (!showTrackSegments) {
       return pointsLayers
@@ -145,7 +141,7 @@ class VesselsEventsGenerator {
           visibility: 'visible',
         },
         paint: {
-          'line-color': 'white',
+          'line-color': ['get', 'color'],
           'line-width': [...activeFilter, 6, 1.5],
           'line-opacity': 1,
         },
@@ -165,6 +161,7 @@ class VesselsEventsGenerator {
       getVesselEventsGeojson: memoizeOne(getVesselEventsGeojson),
       getVesselSegmentsGeojson: memoizeOne(getVesselSegmentsGeojson),
       filterGeojsonByTimerange: memoizeOne(filterGeojsonByTimerange),
+      filterFeaturesByTimerange: memoizeOne(filterFeaturesByTimerange),
     })
 
     return {