+++ packages/layer-composer/src/generators/vessel-events/vessel-events.utils.ts
@@ -1,5 +1,5 @@
 import { DateTime } from 'luxon'
-import { FeatureCollection } from 'geojson'
+import { Feature, FeatureCollection } from 'geojson'
 import { Segment, segmentsToGeoJSON } from '@globalfishingwatch/data-transforms'
 import { Dictionary } from '../../types'
 import filterTrackByTimerange from '../track/filterTrackByTimerange'
@@ -25,27 +25,6 @@ const getEncounterAuthColor = (authorizationStatus: AuthorizationOptions) => {
     default:
       return ''
   }
-}
-
-export const setActiveEvent = (
-  data: FeatureCollection,
-  currentEventId: string | null
-): FeatureCollection => {
-  const featureCollection = { ...data }
-  featureCollection.features = featureCollection.features.map((feature) => {
-    const newFeature = { ...feature }
-    newFeature.properties = newFeature.properties || {}
-    newFeature.properties.active = currentEventId
-      ? newFeature.properties.id === currentEventId
-      : false
-    return newFeature
-  })
-  featureCollection.features.sort((a, b) => {
-    if (a.properties && a.properties.active) return 1
-    else if (b.properties && b.properties.active) return -1
-    else return 0
-  })
-  return featureCollection
 }
 
 const getDateTimeDate = (date: string | number) => {
@@ -94,6 +73,21 @@ export const getVesselEventsGeojson = (trackEvents: RawEvent[] | null): FeatureC
   return featureCollection
 }
 
+export const filterFeaturesByTimerange = (features: Feature[], start: string, end: string) => {
+  if (start && end) {
+    const startMs = new Date(start).getTime()
+    const endMs = new Date(end).getTime()
+    return features.filter((feature) => {
+      return (
+        feature.properties &&
+        feature.properties.timestamp > startMs &&
+        feature.properties.timestamp < endMs
+      )
+    })
+  }
+  return features
+}
+
 export const filterGeojsonByTimerange = (
   geojson: FeatureCollection,
   start: string,
@@ -133,19 +127,28 @@ export const getVesselSegmentsGeojson = (track: any, data: RawEvent[]): FeatureC
     : segmentsToGeoJSON(track as Segment[])
   if (!geojson) return featureCollection
   featureCollection.features = data.flatMap((event: RawEvent) => {
-    return filterTrackByTimerange(geojson, event.start, event.end).features.map((feature) => ({
-      ...feature,
-      properties: {
-        id: event.id,
-        type: event.type,
-        start: getDateTimeDate(event.start).toUTC().toISO(),
-        end: getDateTimeDate(event.end).toUTC().toISO(),
-        ...(event.vessel && {
-          vesselId: event.vessel.id,
-          vesselName: event.vessel.name,
-        }),
-      },
-    }))
+    return filterTrackByTimerange(geojson, event.start, event.end).features.map((feature) => {
+      const authorizationStatus = event.encounter
+        ? event.encounter.authorizationStatus
+        : ('unmatched' as AuthorizationOptions)
+      return {
+        ...feature,
+        properties: {
+          id: event.id,
+          type: event.type,
+          start: getDateTimeDate(event.start).toUTC().toISO(),
+          end: getDateTimeDate(event.end).toUTC().toISO(),
+          color:
+            event.type === 'encounter'
+              ? getEncounterAuthColor(authorizationStatus)
+              : EVENTS_COLORS[event.type],
+          ...(event.vessel && {
+            vesselId: event.vessel.id,
+            vesselName: event.vessel.name,
+          }),
+        },
+      }
+    })
   })
   return featureCollection
 }